feat(dash): new UI for home
