+++ apps/web/src/app/page.tsx
@@ -1,48 +1,88 @@
-'use client'
-import { graphql, useLazyLoadQuery } from "react-relay";
-import { useEffect, useState } from "react";
-import Menu from "@/components/Menu";
-import Product from "@/components/Product";
-import type { pageQuery, pageQuery$data } from "./__generated__/pageQuery.graphql";
+import Link from "next/link"
+import { Home, Package2Icon, PackageSearch, Settings, ShoppingCart, UsersRound } from "lucide-react"
+import ProductsDashboard from "@/components/ProductsDashboard"
 
-export default function Page(): JSX.Element {
-  const response = useLazyLoadQuery<pageQuery>(
-    graphql`
-      query pageQuery {
-        products(first: 100)
-          @connection(key: "ProductList_products") {
-          edges {
-            cursor
-            node {
-              ...ProductFragment_product
-            }
-          }
-        }
-      }
-    `,
-    {},
-    { fetchPolicy: 'network-only' }
-  )
-
-  const [products, setProducts] = useState<pageQuery$data>()
-
-  useEffect(() => {
-    setProducts(response)
-  }, [response])
-
+export default function Component() {
   return (
-    <>
-      <Menu />
-      <div className="mt-28 flex flex-col sm:flex-row justify-center gap-4">
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-3">
-          {products?.products.edges?.map((edge) => {
-            if (!edge?.node) return null
-            return (
-              <Product key={edge.cursor} product={edge.node} />
-            )
-          })}
+    <div className="grid min-h-screen w-full overflow-hidden lg:grid-cols-[280px_1fr]">
+      <div className="hidden border-r bg-gray-100/40 lg:block dark:bg-gray-800/40">
+        <div className="flex flex-col gap-2">
+          <div className="flex h-[60px] items-center px-6">
+            <Link className="flex items-center gap-2 font-semibold" href="#">
+              <Package2Icon className="h-6 w-6" />
+              <span className="">Wecommerce Inc</span>
+            </Link>
+          </div>
+          <div className="flex-1 overflow-auto">
+            <nav className="grid items-start px-4 text-sm font-medium">
+              <Link
+                className="flex items-center gap-3 rounded-lg px-3 py-2 bg-gray-100 text-gray-900   transition-all hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-50"
+                href="#"
+              >
+                <Home className="h-4 w-4" />
+                Home
+              </Link>
+              <Link
+                className="flex items-center gap-3 rounded-lg  px-3 py-2 text-gray-500  transition-all hover:text-gray-900 dark:bg-gray-800 dark:text-gray-50 dark:hover:text-gray-50"
+                href="#"
+              >
+                <PackageSearch className="h-4 w-4" />
+                Products - Coming soon
+              </Link>
+              <Link
+                className="flex items-center gap-3 rounded-lg px-3 py-2 text-gray-500 transition-all hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-50"
+                href="#"
+              >
+                <ShoppingCart className="h-4 w-4" />
+                Orders - Coming soon
+              </Link>
+              <Link
+                className="flex items-center gap-3 rounded-lg px-3 py-2 text-gray-500 transition-all hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-50"
+                href="#"
+              >
+                <UsersRound className="h-4 w-4" />
+                Customers - Coming soon
+              </Link>
+              <Link
+                className="flex items-center gap-3 rounded-lg px-3 py-2 text-gray-500 transition-all hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-50"
+                href="#"
+              >
+                <Settings className="h-4 w-4" />
+                Settings
+              </Link>
+            </nav>
+          </div>
         </div>
       </div>
-    </>
-  );
+      <div className="flex flex-col">
+        <header className="flex h-14 lg:h-[60px] items-center gap-4 border-b bg-gray-100/40 px-6 dark:bg-gray-800/40">
+          <Link className="lg:hidden" href="#">
+            <Package2Icon className="h-6 w-6" />
+            <span className="sr-only">Home</span>
+          </Link>
+          <div className="flex-1">
+            <h1 className="font-semibold text-lg">Products</h1>
+          </div>
+          <div className="flex flex-1 items-center gap-4 md:ml-auto md:gap-2 lg:gap-4">
+            {/* <form className="ml-auto flex-1 sm:flex-initial">
+              <div className="relative">
+                <SearchIcon className="absolute left-2.5 top-2.5 h-4 w-4 text-gray-500 dark:text-gray-400" />
+                <Input
+                  className="pl-8 sm:w-[300px] md:w-[200px] lg:w-[300px] bg-white"
+                  placeholder="Search orders..."
+                  type="search"
+                />
+              </div>
+            </form> */}
+          </div>
+        </header>
+        <main className="flex flex-1 flex-col gap-4 p-4 md:gap-8 md:p-6">
+          <div className="p-2">
+            <ProductsDashboard />
+          </div>
+        </main>
+      </div>
+    </div>
+  )
 }
+
